Move intcode computer to its own file and class
