+++ advent_of_code_2019_python/__init__.py
@@ -1,0 +1,3 @@
+from .intcode_computer import IntcodeComputer
+
+__all__ = ['IntcodeComputer']
